% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{build.arima}
\alias{build.arima}
\title{Builds an ARIMA-based parsnip model}
\usage{
build.arima(engine)
}
\arguments{
\item{engine}{A string storing the parsnip engine the model
is intended to run on}
}
\value{
A workflow object storing the un-trained model
}
\description{
Builds and encapsulates sequential models (ARIMA, ARIMA with
Boosted Errors, Auto-ARIMA, Auto-ARIMA with Boosted Errors)
inside a workflow. This model is untrained and simply stores
the model architecture and specifications. Currently supported
engines are: 'arima', 'auto_arima', 'arima_xgboost',
'auto_arima_xgboost'
}
\examples{
build.arima('arima')
}
