% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{build.tree}
\alias{build.tree}
\title{Builds a tree-based parsnip model}
\usage{
build.tree(engine)
}
\arguments{
\item{engine}{A string storing the parsnip engine the model
is intended to run on}
}
\value{
A workflow object storing the un-trained model
}
\description{
Builds and encapsulates tree-based models (Random Forests,
Decision Trees, etc.) inside a workflow. This model is untrained
and simply stores the model architecture and specifications.
Currently supported engines are: 'ranger', 'randomForest', 'rpart'
and 'spark'
}
\examples{
build.tree('ranger')
}
