% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{BuildForecasts}
\alias{BuildForecasts}
\title{Generates ensembles and final forecasts from trained models}
\usage{
BuildForecasts(
  models,
  prepared_tbl,
  future_tbl,
  valid_data,
  ensembles = c("mean", "median", "rank"),
  metric = "rmse"
)
}
\arguments{
\item{models}{A list of trained parsnip models, each encompassed in a
workflow}

\item{prepared_tbl}{The full processed dataset (both training and validation
datasets)}

\item{future_tbl}{The empty frame storing the horizon data to forecast}

\item{valid_data}{The dataset on which to evaluate models}

\item{ensembles}{A vector storing the ensembles to be generated. Possible
options are "mean", "median", or rank"}

\item{metric}{The metric by which to rank models (e.g. rmse, mape, mse, etc.)}
}
\value{
A final table storing the models' forecasts on the horizon data
}
\description{
Generates ensembles and final forecasts from trained models
}
