% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{BuildRecipe}
\alias{BuildRecipe}
\title{Compile pre-processing recipe for training dataset}
\usage{
BuildRecipe(train_data)
}
\arguments{
\item{train_data}{The dataset to be used for model training}
}
\value{
A recipe definition containing pre-processing steps
}
\description{
Compile pre-processing recipe for training dataset
}
