% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{BuildTemporal}
\alias{BuildTemporal}
\title{Formats the temporal column of a given tibble}
\usage{
BuildTemporal(data, origin = NULL)
}
\arguments{
\item{data}{The tibble created in load.data}

\item{origin}{Optional parameter specifying the earliest point at which data
in the tibble originates from. This parameter is used when the
temporal column has a numeric type (each number is assumed to
be the number of time units since the origin, and is hence
added to the origin to return a column of Dates)}
}
\value{
A new dataframe with a processed temporal column (of type 'Date')
}
\description{
Formats the temporal column of a given tibble
}
