% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{check.data}
\alias{check.data}
\title{Helper function to ensure user-provided dataset is formatted correctly}
\usage{
check.data(data)
}
\arguments{
\item{data}{Dataset built by user}
}
\value{
TRUE if the dataset is correctly formatted - throw an error otherwise
}
\description{
Helper function to ensure user-provided dataset is formatted correctly
}
