% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{CompileModels}
\alias{CompileModels}
\title{Compiles a collection of trained models into a single list}
\usage{
CompileModels(
  include = c(),
  exclude = c(),
  seasonalities,
  recipe_spec,
  train_data
)
}
\arguments{
\item{include}{A list of engines to build models around. Should only be
specified if exclude is empty. If both include and exclude are
empty, all supported engines will be trained}

\item{exclude}{A list of engines to exclude from model training. If specified,
all supported engines EXCEPT those specified here will be
trained. If both 'include' and 'exclude' are empty, all
supported engines will be trained.}

\item{seasonalities}{A vector storing all seasonalities exhibited by the time
series. Possible values are: "daily", "weekly" or
"yearly"}

\item{recipe_spec}{A recipe definition storing list of pre-processing steps
to be applied on training dataset}

\item{train_data}{Dataset on which to train model collection}
}
\value{
A list of trained parsnip models, each encompassed in a workflow
}
\description{
Compiles a collection of trained models into a single list
}
