% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{fit.to.recipe}
\alias{fit.to.recipe}
\title{Fits a model workflow to the training data based on a pre-processing recipe}
\usage{
\method{fit}{to.recipe}(wflw, recipe_spec, train_data)
}
\arguments{
\item{wflw}{The parsnip model to train, encompassed in a workflow}

\item{recipe_spec}{The recipe storing pre-processing steps}

\item{train_data}{The dataset to use for model training}
}
\value{
A fitted parsnip model, encompassed in a workflow
}
\description{
Fits a model workflow to the training data based on a pre-processing recipe
}
