% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{GenerateFeatures}
\alias{GenerateFeatures}
\title{Engineer features in a time series dataset}
\usage{
GenerateFeatures(data)
}
\arguments{
\item{data}{The dataset to engineer}
}
\value{
A list storing two tibbles - one containing the processed dataset
        provided, and another containing a frame of the horizon (the period
        to forecast)
}
\description{
Engineer features in a time series dataset
}
