% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{LoadData}
\alias{LoadData}
\title{Load in a CSV file as a tibble from a specified path}
\usage{
LoadData(data = NULL, path = NULL, origin = NULL)
}
\arguments{
\item{data}{Optional argument (must be provided if 'path' is null) specifying
a pre-loaded dataframe to be processed. If specified, 'path' will
be ignored. The provided tibble will be checked and the relevant
columns will be filtered.}

\item{path}{Optional argument(must be provided if 'data' is null) specifying
a path to a CSV file. Will be ignored unless 'data' is null. The
provided CSV file will be read, and the data will be checked and
filtered.}

\item{origin}{Optional argument specifying the time of the earliest
observation recorded in the dataset. This parameter will only
be used if the temporal column of the dataset is 'numeric'
instead of 'Date'.}
}
\value{
A tibble storing the temporal and target columns from the original
        dataframe.
}
\description{
Load in a CSV file as a tibble from a specified path
}
