% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{PrettifyTimes}
\alias{PrettifyTimes}
\title{Generate a timetk friendly string from an entered period}
\usage{
PrettifyTimes(period, frequency)
}
\arguments{
\item{period}{The number of time units (should be an integer)}

\item{frequency}{The unit each row of data represents (e.g. "day", "year")}
}
\value{
A string combining the period with the time units (e.g. "60 days")
}
\description{
Generate a timetk friendly string from an entered period
}
